Refactor xtask docs command into modular layout
