+++ xtask/src/commands/docs/mod.rs
@@ -1,0 +1,8 @@
+mod cli;
+mod executor;
+mod validation;
+
+#[allow(unused_imports)]
+pub use cli::usage;
+pub use cli::{DocsOptions, parse_args};
+pub use executor::execute;
